+++ monument/cli/src/spec.rs
@@ -10,7 +10,7 @@ use std::{
 use bellframe::{
     method::LABEL_LEAD_END,
     method_lib::QueryError,
-    music::{Regex, RegexElem},
+    music::{Elem, Pattern},
     place_not::PnBlockParseError,
     Bell, Mask, MethodLib, Row, RowBuf, Stage, Stroke,
 };
@@ -450,12 +450,10 @@ impl Spec {
                 let left_bell = right_bell + 1;
                 // For every handbell pair, we need patterns for `*<left><right>` and `*<right><left>`
                 for (b1, b2) in [(left_bell, right_bell), (right_bell, left_bell)] {
-                    let regex = Regex::from_elems(
-                        [RegexElem::Glob, RegexElem::Bell(b1), RegexElem::Bell(b2)],
-                        stage,
-                    );
-                    let mask =
-                        Mask::from_regex(&regex).expect("Handbell patterns should always be valid");
+                    let pattern =
+                        Pattern::from_elems([Elem::Star, Elem::Bell(b1), Elem::Bell(b2)], stage);
+                    let mask = Mask::from_pattern(&pattern)
+                        .expect("Handbell patterns should always be valid");
                     add_ch_pattern(&mask, self.handbell_coursing_weight);
                 }
             }
@@ -732,7 +730,7 @@ fn mask_parse_error(
     use bellframe::mask::ParseError as PE;
     let mut s = format!("Can't parse {} {:?}: ", mask_kind, string);
     match e {
-        PE::MultipleGlobs => {
+        PE::MultipleStars => {
             s.push_str("too many `*`s.  Masks can only have one region with `x` or `*`.");
         }
         PE::BellExceedsStage(bell, stage) => {